// src/app/Models/Sales/ClienteResponse.model.ts
export interface ClienteResponseModel {
  apellidoMaterno: string;
  apellidoPaterno: string;
  calle: string;
  ciudad: string;
  codigoCliente: number;
  codigoPostal: string;
  codigoTienda: number;
  email: string;
  estado: string;
  noExterior: string;
  noInterior: string;
  nombre: string;
  telefono: string;
  codigoCaja: number;
}